requirements.txt
python-dotenv~=1.0.1
Flask~=3.1.0